enlaces a página de logros actualizada
