+++ src/components/pages/index.js
@@ -2,10 +2,9 @@ import Login from './login';
 import Register from './register';
 import MeProfile from './meProfile';
 import MeEditProfile from './meEditProfile';
+import MeLogros from './meLogros';
 import SearchProfile from './searchProfile';
 import UserProfile from './userProfile';
-import Logros from './meLogros';
-
 import FrequentQuestion from './frequentQuestion';
 import PrivacyConsumers from './privacyConsumers';
 
@@ -16,7 +15,7 @@ export {
   MeEditProfile,
   SearchProfile,
   UserProfile,
-  Logros,
+  MeLogros,
   FrequentQuestion,
   PrivacyConsumers,
 };
