+++ src/components/pages/meLogros/view.js
@@ -23,7 +23,7 @@ import DeleteLogros from '../../modals/deleteLogros';
 const Transition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
 });
-export default function Logros() {
+export default function MeLogros() {
   const [userData, setUserData] = useState({});
 
   useEffect(() => {
